Move apiHook
