+++ nvdaHelper/common/apiHook.h
