+++ nvdaHelper/remote/gdiHooks.cpp
@@ -22,7 +22,7 @@ http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 #include <usp10.h>
 #include "nvdaHelperRemote.h"
 #include "dllmain.h"
-#include "apiHook.h"
+#include <common/apiHook.h>
 #include "displayModel.h"
 #include <common/log.h>
 #include <remote/nvdaControllerInternal.h>
