+++ nvdaHelper/remote/sconscript
@@ -86,7 +86,7 @@ source = [
 	"injection.cpp",
 	"log.cpp",
 	"inProcess.cpp",
-	"apiHook.cpp",
+	"../common/apiHook.cpp",
 	"inputLangChange.cpp",
 	"typedCharacter.cpp",
 	"ime.cpp",
